requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
ecdsa==0.19.1
fastapi==0.128.0
h11==0.16.0
idna==3.11
iso8601==2.1.0
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
pypika-tortoise==0.6.3
python-dotenv==1.2.1
python-jose==3.5.0
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
starlette==0.50.0
tortoise-orm==0.25.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
